feat(mlop-2760): renaming classes and file
